refactor flux_support
