+++ flux_support/src/extern_specs/mod.rs
@@ -1,3 +1,5 @@
 mod iter;
 mod option;
 mod slice;
+mod non_null;
+mod num;
