+++ flux_support/src/extern_specs/non_null.rs
@@ -1,0 +1,3 @@
+#[flux_rs::extern_spec(core::ptr)]
+#[refined_by(n: int)]
+struct NonNull<T>;
